login, signup, logout, logincheck, mainpage, privatepage - routes finished
