+++ routes/auth.js
@@ -63,7 +63,7 @@ router.post('/login', (req, res, next) => {
           return res.render('/', {errorMessage: "Passwort ist falsch"});
       
       req.session.user = foundUser;
-      res.render('auth/home-private');
+      res.render('auth/home-main');
       });
   })
   .catch(err => next (err));
