+++ views/auth/home-main.hbs
@@ -1,5 +1,20 @@
 <!--- ABOUT APP -->
+<header>
+        <div id="logo">
+                <a href="home-main"><li>MyCarApp</li></a>
+        </div>
 
+        <nav id="main-nav">
+            <ul>
+                <a href="home-main"><li>Home</li></a>
+                <a href="home-main"><li>About</li></a>
+                <a href="home-main"><li>Contact</li></a>
+                <a href="myaccount"><li>MyAccount</li></a>
+                <a href="/"><li>Logout</li></a>
+            </ul>
+        </nav>
+    </header>
+    
 <section>
  <div>
    Beschreibung My-Car-App
