+++ views/auth/myaccount.hbs
@@ -2,15 +2,13 @@
 
 <header>
         <div id="logo">
-                <a href="/home-main"><li>MyCarApp</li></a>
+                <a href="home-main"><li>MyCarApp</li></a>
         </div>
 
         <nav id="main-nav">
             <ul>
                 <a href="home-main"><li>Home</li></a>
-                <a href="home-main"><li>About</li></a>
-                <a href="home-main"><li>Contact</li></a>
-                <a href="auth/home-private"><li>MyAccount</li></a>
+                <a href="myaccount"><li>MyAccount</li></a>
                 <a href="/"><li>Logout</li></a>
             </ul>
         </nav>
