+++ views/index.hbs
@@ -58,10 +58,10 @@
   <h2>Anmelden</h2>
 
   
-  <input type="email" name="email" placeholder="E-Mail-Adresse" required><br>
+  <input type="email" name="email" placeholder="E-Mail-Adresse" value="muster@web.de"><br>
 
   
-  <input type="password" name="password" placeholder="**********" required><br>
+  <input type="password" name="password" placeholder="**********" value="mustermuster"><br>
 
   {{#if errorMessage }}
     <div class="error-message">{{ errorMessage }}</div><br>
